make contract hooks have a single instance
