+++ apps/trading/pages/portfolio/deposit/use-approve.ts
@@ -1,6 +1,6 @@
 import { DepositPage_assets } from '@vegaprotocol/graphql';
-import { useEthereumTransaction } from './use-ethereum-transaction';
-import { useTokenContract } from './use-token-contract';
+import { useEthereumTransaction } from '../../../hooks/use-ethereum-transaction';
+import { useTokenContract } from '../../../hooks/use-token-contract';
 
 export const useApprove = (
   bridgeAddress: string,
