--- apps/trading/pages/portfolio/deposit/use-bridge-contract.ts
@@ -1,16 +1,0 @@
-import { VegaErc20Bridge } from '@vegaprotocol/smart-contracts-sdk';
-import { useWeb3React } from '@web3-react/core';
-import { useMemo } from 'react';
-
-export const useBridgeContract = () => {
-  const { provider } = useWeb3React();
-  const contract = useMemo(() => {
-    if (!provider) {
-      return null;
-    }
-    // @ts-ignore TODO: Setup env
-    return new VegaErc20Bridge('TESTNET', provider, provider.getSigner());
-  }, [provider]);
-
-  return contract;
-};
